Rename freshstart to readme.md
Hi Ray, I changed the title of the file from "freshstart" to readme.md, this means github will automatically display it as the readme page for this repo. Try messing around by adding some markdown text to the page now if you like. I'll add some resources to canvas today to help and we will do this in class on thursday! thanks